First Login test
Created a reference from GUIMain to MainController.
This reference is now used in the ActionListener of the loginButton of
LoginSeite.java
Changed the datatype of the password to char[] to fit the getPassword()
method (for security reasons) in the login method of MainController, as
well as the getNutzer method of DBrequest.